[Refactor] Update package of Write (#26)
